+++ tests/test_misc.py
@@ -353,25 +353,6 @@ def test_password_unicode_password_salt(client):
     assert b"Welcome matt@lp.com" in response.data
 
 
-@pytest.mark.filterwarnings(
-    "ignore:.*'unauthorized_handler' has been replaced.*:DeprecationWarning"
-)
-def test_set_unauthorized_handler(app, client):
-    @app.security.unauthorized_handler
-    def unauthorized():
-        app.unauthorized_handler_set = True
-        return "unauthorized-handler-set", 401
-
-    app.unauthorized_handler_set = False
-
-    authenticate(client, "joe@lp.com")
-    response = client.get("/admin", follow_redirects=True)
-
-    assert app.unauthorized_handler_set is True
-    assert b"unauthorized-handler-set" in response.data
-    assert response.status_code == 401
-
-
 @pytest.mark.registerable()
 def test_custom_forms_via_config(app, sqlalchemy_datastore):
     class MyLoginForm(LoginForm):